0-Python/final-project/store_locator/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
psycopg2-binary==2.9.9
redis==5.0.1
geopy==2.4.1
bcrypt==4.1.2
python-dotenv==1.0.0
pandas==2.1.4
pytest==7.4.3